Modify package and file names for protobuf files to prevent conflict.
Prior to this change, newer versions of protobuf would display
the following error when the code was running:

"
2020/06/27 04:54:35 WARNING: proto: file "protobuf.proto" is already registered
A future release will panic on registration conflicts. See:
https://developers.google.com/protocol-buffers/docs/reference/go/faq#namespace-conflict
"

This was due to the fact that the sysboxMgrGrpc and sysboxFsGrpc where both
generating a file named "protobuf.proto". This change fixes this by
having each generate files with more specific names.