Fix for race-condition in poll-server code and doc improvements
This one fixes a race-condition triggered in the scenario where both a 'poll-delete' request is simultaneously processed along a 'poll-add' arrival. This is typically the case when
a syscall is trapped, and simultaneously, its underlying seccomp-bpf FD is closed.

There were two bugs in this logic:

* Upon arrival of a poll-delete action, we were not fully protecting the share-state associated with the pollAction object itself, which can be potentially consumed by the 'poll-add' logic.
The fix for this one is to simply expand the critical section of the poll-delete's request logic.

* Upon arrival of a poll-add action, we were blindly pushing the matching pollAction object found in the pollServer DB, without first verifying that this one corresponded to a poll-addition
request. In this scenario we could end up injecting a 'poll-delete' request into the poll-server's event-loop instead of the intended 'poll-add' one. The poll-server fsm would then discard
this request if, by then, the poll-fd had been eliminated (as part of the 'poll-delete' processing), which would prevent this one from being able to send back an 'awake' message to the
poll-server client. Technically speaking, this wouldn't impact the user as the seccomp-FD is going down anyways, but this one could potentially affect sysbox-fs performance numbers as we would
be leaving dangling tracer goroutines behind.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>