Fix pollserver's race-condition issue involving concurrent add/delete ops
This PR fixes a race-condition issue observed in scenarios where a pollServer client concurrently receives ingress traffic as well as a 'delete' message. Within the context of the sysbox-fs' syscall traffic handling, this can happen when, for example, a 'mount' syscall is received
from the seccomp-tracee, and at the same time (or right after), this seccomp-tracee process is killed (which would be notified by the sysbox-fs' pidMonitor).

In this scenario, the current code could end up missing the "seccomp-fd delete" event, which could potentially translate into a goroutine leak issue (i.e. connHandler), or even a 'stalling' situation.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>