Increase sysbox-mgr grpc timeout due to rootfs chown delays.
For hosts without shiftfs, Sysbox performs chown of the rootfs.
To do this quickly, the sysbox-mgr clones the rootfs and
enables metacopy=on on the rootfs overlayfs mount.

However, even with metacopy=on, the rootfs chown can take a
good amount of time. And if several containers are started
at the same time, the delay can add up causing the grpc
channel between sysbox-runc and sysbox-mgr to time out.

This commit fixes this by increasing the timeout from
20 secs to 30 secs. Tests verify that this is sufficient.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>