Increase sysbox-runc -> sysbox-mgr gRPC timeout.
Back in commit 4da2d58, we the timeout to 30 secs, which is quite large already.

However, the sysbox performance tests would fail sometimes because they hit this
limit when Sysbox runs without shiftfs and therefore has to clone to rootfs and
chown it. The cloning is very fast, but the chown can take longer, particularly
on large container images, even though it's being done when the rootfs is using
overlayfs with metacopy=on.

Here we further increase the timeout to 40 secs. With this change, the Sysbox
performance tests consistently pass.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>